converted all logging to compile time static to save on space, refactored program entrypoints, fixed batching logic, wip testing and making initializations log conditional
