+++ src/recurrent/myrecords.py
@@ -22,16 +22,17 @@ class OhoData[Data](eqx.Module):
 class GodState(eqx.Module):
     prng: PRNG
     start_epoch: int
-    start_batch: int
+    start_example: int
     innerTimeConstant: float = eqx.field(static=True)
     outerTimeConstant: float = eqx.field(static=True)
+    globalLogConfig: GlobalLogConfig = eqx.field(static=True)
 
     # Inner fields
     rnnState: Optional[RnnState] = eqx.field(default=None)
     innerInfluenceTensor: Optional[JACOBIAN] = eqx.field(default=None)
     innerUoro: Optional[UORO_Param] = eqx.field(default=None)
     innerLogs: Optional[Logs] = eqx.field(default=None)
-    innerLogConfig: Optional[LogConfig] = eqx.field(default=None)
+    innerLogConfig: Optional[LogConfig] = eqx.field(default=None, static=True)
     innerOptState: Optional[optax.OptState] = eqx.field(default=None)
     innerSgdParameter: Optional[SgdParameter] = eqx.field(default=None)
     innerAdamParameter: Optional[AdamParameter] = eqx.field(default=None)
@@ -40,7 +41,7 @@ class GodState(eqx.Module):
     outerInfluenceTensor: Optional[JACOBIAN] = eqx.field(default=None)
     outerUoro: Optional[UORO_Param] = eqx.field(default=None)
     outerLogs: Optional[Logs] = eqx.field(default=None)
-    outerLogConfig: Optional[LogConfig] = eqx.field(default=None)
+    outerLogConfig: Optional[LogConfig] = eqx.field(default=None, static=True)
     outerOptState: Optional[optax.OptState] = eqx.field(default=None)
     outerSgdParameter: Optional[SgdParameter] = eqx.field(default=None)
     outerAdamParameter: Optional[AdamParameter] = eqx.field(default=None)
@@ -64,6 +65,7 @@ class GodInterpreter:
     getRnnConfig: LocalApp[RnnConfig]
     getTimeConstant: LocalApp[float]
     getLogConfig: LocalApp[LogConfig]
+    getGlobalLogConfig: LocalApp[GlobalLogConfig]
     putLogs: Callable[[Logs], LocalApp[Unit]]
 
     getRnnParameter: LocalApp[RnnParameter]
@@ -79,9 +81,6 @@ class GodInterpreter:
         prng, new_prng = yield from gets(lambda e: prng_split(e.prng))
         _ = yield from modifies(lambda e: eqx.tree_at(lambda t: t.prng, e, new_prng))
         return pure(prng, PX[tuple[Self, GodState]]())
-
-
-test = "Hi"
 
 
 @dataclass(frozen=True)
@@ -126,30 +125,3 @@ class GodConfig:
     outer_clip_sharpness: float
     inner_log_expensive: Optional[bool] = None
     outer_log_expensive: Optional[bool] = None
-
-
-# def rnn_array(state: State) -> jax.Array:
-#     return toVector(endowVector(state.activation))
-
-
-# def to_rnn_array(state: State, rec_state: jax.Array) -> State:
-#     activation = invmap(state.activation, lambda _: rec_state)
-#     return eqx.tree_at(lambda t: t.activation, state, activation)
-
-
-# def parameter_array(state: State) -> jax.Array:
-#     return toVector(endowVector(state.rnnParameter))
-
-
-# def to_parameter_array(state: State, rec_param: jax.Array) -> State:
-#     rnnParameter = invmap(state.rnnParameter, lambda _: rec_param)
-#     return eqx.tree_at(lambda t: t.rnnParameter, state, rnnParameter)
-
-
-# def sgd_array(state: State) -> jax.Array:
-#     return toVector(endowVector(state.sgdParameter))
-
-
-# def to_sgd_array(state: State, sgd_param: jax.Array) -> State:
-#     sgdParameter = invmap(state.sgdParameter, lambda _: sgd_param)
-#     return eqx.tree_at(lambda t: t.sgdParameter, state, sgdParameter)
